# application/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/city": {
            "get": {
                "tags": [
                    "City"
                ],
                "summary": "Retrieves city models",
                "operationId": "7363283289731a0776ed7f6bd67cbcfa",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "City list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/City"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": [
                                            {
                                                "id": 1,
                                                "cityName": "Новосибирск"
                                            },
                                            {
                                                "id": 2,
                                                "cityName": "Москва"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "City"
                ],
                "summary": "Adds new city",
                "operationId": "d7744ae6571c47ec6cb646d19d6e9696",
                "parameters": [
                    {
                        "name": "cityName",
                        "in": "query",
                        "description": "New city name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New city",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                },
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "cityName": "Новосибирск"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/job": {
            "get": {
                "tags": [
                    "Job"
                ],
                "summary": "Retrieves job models",
                "operationId": "07aee1de1490cc4cd8530a3300610d50",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search string",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Job"
                                            }
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "data": [
                                            {
                                                "id": 1,
                                                "jobTitle": "Специалист по организации перевозок"
                                            },
                                            {
                                                "id": 2,
                                                "jobTitle": "Специалист по грузоперевозкам"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Job"
                ],
                "summary": "Adds new job",
                "operationId": "69c15ec33d635f6de9628e023432e17d",
                "parameters": [
                    {
                        "name": "jobTitle",
                        "in": "query",
                        "description": "New job name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "New job",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                },
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "jobTitle": "Координатор по логистике"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/route": {},
        "/api/city/{id}": {
            "delete": {
                "tags": [
                    "City"
                ],
                "summary": "Deletes the city",
                "description": "Remove the specified resource from storage.",
                "operationId": "6e1e0ab1861a1b9f6db499dbac60ffb0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the city to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "tags": [
                    "City"
                ],
                "summary": "Updates the city",
                "description": "Update the specified resource in storage.",
                "operationId": "9d411fcd44383d268923b54f963344c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the city to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer"
                        }
                    },
                    {
                        "name": "cityName",
                        "in": "query",
                        "description": "New city name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated city",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/City"
                                },
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "cityName": "Новосибирск"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/job/{id}": {
            "delete": {
                "tags": [
                    "Job"
                ],
                "summary": "Deletes the job",
                "description": "Remove the specified resource from storage.",
                "operationId": "28fa567d616ec3a50685b09976006092",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                }
            },
            "patch": {
                "tags": [
                    "Job"
                ],
                "summary": "Updates the job",
                "description": "Update the specified resource in storage.",
                "operationId": "6ca600607755a475054c9a5621176ad1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the job to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "integer"
                        }
                    },
                    {
                        "name": "jobTitle",
                        "in": "query",
                        "description": "New job name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated job",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                },
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "jobTitle": "Координатор по логистике"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "City": {
                "title": "City",
                "properties": {
                    "id": {
                        "description": "City id",
                        "type": "integer"
                    },
                    "cityName": {
                        "description": "City name",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "Job": {
                "title": "Job",
                "properties": {
                    "id": {
                        "description": "Job id",
                        "type": "integer"
                    },
                    "jobTitle": {
                        "description": "Job title",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}